\name{ShapeFiles}
\alias{CAN_PROV_ShapeFile}
\alias{MB_DA_ShapeFile}
\alias{MB_RHAD_ShapeFile}
\docType{data}
\title{
Useful Shape Files
}
\description{
Useful shape files for data analysis and mapping.
}
\usage{data("CAN_PROV_ShapeFile")
data("MB_DA_ShapeFile")
data("MB_RHAD_ShapeFile")}
\format{

}
\details{
CAN_PROV_ShapeFile provides the polygons for the Canadian Provinces.

MB_DA_ShapeFile provides the polygons for the Manitoba Dissemination Areas.

MB_RHAD_ShapeFile provides the polygons for the Manitoba Regional Health Authorioty Districts.

}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
prov_shape <- CAN_PROV_ShapeFile

}
\keyword{datasets}
